Moves the methods from the product_helper to the yoast_helper
