+++ admin/class-admin-init.php
@@ -353,7 +353,7 @@ class WPSEO_Admin_Init {
 	 * @return void
 	 */
 	private function register_premium_upsell_admin_block() {
-		if ( ! YoastSEO()->helpers->product->is_premium() ) {
+		if ( ! YoastSEO()->helpers->yoast->is_premium() ) {
 			$upsell_block = new WPSEO_Premium_Upsell_Admin_Block( 'wpseo_admin_promo_footer' );
 			$upsell_block->register_hooks();
 		}
