+++ admin/class-config.php
@@ -157,7 +157,7 @@ class WPSEO_Admin_Pages {
 	private function should_show_local_seo_upsell() {
 		$addon_manager = new WPSEO_Addon_Manager();
 
-		return ! YoastSEO()->helpers->product->is_premium()
+		return ! YoastSEO()->helpers->yoast->is_premium()
 			&& ! ( defined( 'WPSEO_LOCAL_FILE' ) );
 	}
