+++ admin/metabox/class-metabox.php
@@ -1091,7 +1091,7 @@ class WPSEO_Metabox extends WPSEO_Meta {
 	protected function get_product_title() {
 		$product_title = 'Yoast SEO';
 
-		if ( YoastSEO()->helpers->product->is_premium() ) {
+		if ( YoastSEO()->helpers->yoast->is_premium() ) {
 			$product_title .= ' Premium';
 		}
