+++ admin/taxonomy/class-taxonomy-metabox.php
@@ -164,7 +164,7 @@ class WPSEO_Taxonomy_Metabox {
 	protected function get_product_title() {
 		$product_title = 'Yoast SEO';
 
-		if ( YoastSEO()->helpers->product->is_premium() ) {
+		if ( YoastSEO()->helpers->yoast->is_premium() ) {
 			$product_title .= ' Premium';
 		}
