+++ wp-seo-main.php
@@ -188,7 +188,7 @@ function _wpseo_activate() {
 	}
 
 	// Reset tracking to be disabled by default.
-	if ( ! YoastSEO()->helpers->product->is_premium() ) {
+	if ( ! YoastSEO()->helpers->yoast->is_premium() ) {
 		WPSEO_Options::set( 'tracking', false );
 	}
 
@@ -371,7 +371,7 @@ function wpseo_admin_init() {
  * on PHP 5.3+, the constant should only be set when requirements are met.
  */
 function wpseo_cli_init() {
-	if ( YoastSEO()->helpers->product->is_premium() ) {
+	if ( YoastSEO()->helpers->yoast->is_premium() ) {
 		WP_CLI::add_command(
 			'yoast redirect list',
 			'WPSEO_CLI_Redirect_List_Command',
